requirements.txt
Flask==2.0.3
requests==2.27.1
scikit-learn==1.0.2
numpy==1.21.5
tqdm==4.63.0
rich==12.0.0

# new dependencies
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113 # why was it 1.9?

pytorch-lightning==1.6.4
jupyterlab
transformers
wandb
jupyterlab
pytorch-pretrained-bert
termcolor
pytorch-crf